Changed some datatypes from double to float.

+++ src/kernel/espmm.hpp
@@ -21,7 +21,7 @@ void equivariant_spmm_cpu(
         uint64_t edge_count,
         uint64_t* rows,
         uint64_t* cols,
-        double* X_in,
-        double* X_out,
-        double* edge_features);
+        float* X_in,
+        float* X_out,
+        float* edge_features);
